use size_t instead of unsigned int, and added insertion sort with tests

+++ src/core/array_ops.c
@@ -1,7 +1,7 @@
 #include"c_algorithms.h"
 
 // Returns 1 when every element has the same value at each index in both arrays, the arrays must be the same length
-int compareParallelArrs(int * arr_a, int * arr_b, unsigned int size){
+int compareParallelArrs(int * arr_a, int * arr_b, size_t size){
     for(int i = 0; i < size; i++){
         if (arr_a[i] != arr_b[i]){
             return -1;
